Moved proto files to their own repo. Pulled in as a submodule.
