+++ .idea/.idea.OddDotNet/.idea/protoeditor.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProtobufLanguageSettings">
+    <option name="importPathEntries">
+      <list>
+        <ImportPathEntry>
+          <option name="location" value="file://$PROJECT_DIR$/src/OddDotNet/OddDotProto" />
+        </ImportPathEntry>
+        <ImportPathEntry>
+          <option name="location" value="file://$PROJECT_DIR$/src/OddDotNet/OpenTelemetry" />
+        </ImportPathEntry>
+      </list>
+    </option>
+  </component>
+</project>
