+++ .idea/.idea.OddDotNet/.idea/vcs.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/src/OddDotNet/OddDotProto" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/src/OddDotNet/OpenTelemetry" vcs="Git" />
   </component>
 </project>
