+++ src/OddDotNet/Proto/Spans/V1/SpanQueryRequest.cs
@@ -1,4 +1,4 @@
-namespace OddDotNet;
+namespace OddDotNet.Proto.Spans.V1;
 
 public sealed partial class SpanQueryRequest : IQueryRequest<Span>
 {
