+++ src/OddDotNet/Proto/Spans/V1/SpanResetRequest.cs
@@ -1,4 +1,4 @@
-namespace OddDotNet;
+namespace OddDotNet.Proto.Spans.V1;
 
 public sealed partial class SpanResetRequest : IResetRequest<Span>
 {
