--- src/OddDotNet/Protos/span_reset_request.proto
@@ -1,5 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "OddDotNet";
-
-message SpanResetRequest{}
