+++ src/OddDotNet/SpanChannelManager.cs
@@ -1,4 +1,5 @@
 using System.Threading.Channels;
+using OddDotNet.Proto.Spans.V1;
 
 namespace OddDotNet;
