+++ tests/OddDotNet.Aspire.Tests/AspireFixture.cs
@@ -1,5 +1,6 @@
 using Grpc.Net.Client;
 using Microsoft.Extensions.DependencyInjection;
+using OddDotNet.Proto.Spans.V1;
 using OpenTelemetry.Proto.Collector.Trace.V1;
 
 namespace OddDotNet.Aspire.Tests;
