+++ tests/OddDotNet.Aspire.Tests/SpanByteStringQueryTests.cs
@@ -1,4 +1,5 @@
 using Google.Protobuf;
+using OddDotNet.Proto.Spans.V1;
 
 namespace OddDotNet.Aspire.Tests;
