+++ tests/OddDotNet.Aspire.Tests/SpanDoubleQueryTests.cs
@@ -1,3 +1,5 @@
+using OddDotNet.Proto.Spans.V1;
+
 namespace OddDotNet.Aspire.Tests;
 
 public class SpanDoubleQueryTests : IClassFixture<AspireFixture>, IAsyncLifetime
