+++ tests/OddDotNet.Aspire.Tests/SpanUInt64QueryTests.cs
@@ -1,3 +1,5 @@
+using OddDotNet.Proto.Spans.V1;
+
 namespace OddDotNet.Aspire.Tests;
 
 public class SpanUInt64QueryTests : IClassFixture<AspireFixture>, IAsyncLifetime
